Refactor FetchSummaryTableTool to use Group Argument
The previous FetchSummaryTable tool supported variables and groups. This created nested, unclear expectations: either group or variables are required, and, in both instances, either for or ucgid are required. This update differentiates fetching data by groups and fetching data by variables to encourage use of the groups tool for larger queries.

* Rename FetchSummaryTableTool to FetchTableByGroupTool
* Add group requirement for FetchTableByGroupTool
* Add custom validation that returns an error when for and ucgid are undefined
* Add custom validation that returns an error when for and ucgid are defined
* Update schemas and tests to reflect the new tool name and functionality
* Update index to import the new tool
* Add Validator Tests for Geography Arguments