Revert FetchSummaryTableTool and Introduce Get Argument
The previous commit removed the ability to query tables by variables - opting to separate group and variable querying. However, variables can be used in conjunction with groups. This commit takes an alternative approach to the required arguments and validation challenges by introduce a get argument with nested variables array and group string, both of which are optional. This ensures either variables or group is passed to get, but individually each is optional - reducing the complexity of validating these inputs.

* Rename FetchTableByGroupTool to FetchSummaryTable
* Add get argument with nested variables array and group string
* Update README to include get example
* Add Variable Limitation Max to Property Description